policy: delete compiler module
