+++ src/policy/mod.rs
@@ -26,7 +26,6 @@
 //!
 
 mod ast;
-mod compiler;
 pub mod descriptor;
 mod embed;
 mod error;
